requirements.txt
# requirements.txt — TapTap Analytics Chatbot v3

# Backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
asyncpg>=0.29.0
python-dotenv>=1.0.1
pydantic>=2.7.4

# LangGraph + LangChain
langgraph>=0.2.14
langchain-core>=0.2.38
langchain-openai>=0.1.23
openai>=1.45.0

# Frontend
streamlit>=1.38.0
requests>=2.32.3
pandas>=2.2.2